requirements.txt
python-dotenv==1.0.1
requests==2.31.0
beautifulsoup4==4.12.2
moviepy==1.0.3
pillow==10.0.0
google-api-python-client==2.91.0
google-auth-oauthlib==1.0.0
ratelimit==2.2.1
tqdm==4.66.1